msm: thermal: Use module kobj and attribute for top-level sysfs
The top-level sysfs parameters are accessed and written using
module_attr_(show|store) instead of kobj_attr_(show|store) that other
module use. This is because the other sysfs is wrapped inside kobj
subdirectories.

While in normal situation this can run fine as it is, it is not the case
with CFI as they enforced the correct function type to be called or
passed. With that said, we should use module designated struct instead
as this is the correct type to be passed and called.

Signed-off-by: Angga Satya Putra <anggasp@hotmail.com>